# api/openapi.yaml
openapi: 3.1.0
info:
  title: Puerta Abierta API
  version: 1.0.0
  summary: STR operations API for Paraguay-focused owners and operators
  description: |
    REST API contract for multi-tenant short-term rental operations:
    organizations, properties, reservations, tasks, finance, and reporting.
servers:
  - url: https://api.puertaabierta.app/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /channels:
    get:
      tags: [Channels]
      summary: List channels
      parameters:
        - $ref: "#/components/parameters/OrgIdQuery"
      responses:
        "200":
          description: Channel list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Channel"
                required: [data]
    post:
      tags: [Channels]
      summary: Create channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChannelInput"
      responses:
        "201":
          description: Channel created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
  /listings:
    get:
      tags: [Listings]
      summary: List listings
      parameters:
        - $ref: "#/components/parameters/OrgIdQuery"
        - name: unit_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Listing list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Listing"
                required: [data]
    post:
      tags: [Listings]
      summary: Create listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingInput"
      responses:
        "201":
          description: Listing created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /listings/{listing_id}/sync-ical:
    post:
      tags: [Listings]
      summary: Trigger iCal sync for listing
      parameters:
        - $ref: "#/components/parameters/ListingIdPath"
      responses:
        "202":
          description: Sync queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [queued]
                required: [status]
...
  /expenses:
    get:
      tags: [Expenses]
      summary: List expenses
      parameters:
        - $ref: "#/components/parameters/OrgIdQuery"
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: category
          in: query
          schema:
            $ref: "#/components/schemas/ExpenseCategory"
      responses:
        "200":
          description: Expense list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Expense"
                required: [data]
    post:
      tags: [Expenses]
      summary: Create expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseInput"
      responses:
        "201":
          description: Expense created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    OrgIdQuery:
      name: org_id
      in: query
      required: true
      schema:
        type: string
        format: uuid
...
    ListingIdPath:
      name: listing_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    UUID:
      type: string
      format: uuid
    Currency:
      type: string
      enum: [PYG, USD]
...
    ChannelKind:
      type: string
      enum: [airbnb, bookingcom, direct, vrbo, other]
...
    ExpenseCategory:
      type: string
      enum: [cleaning, maintenance, utilities, supplies, platform_fee, tax, staff, other]
    PaymentMethod:
      type: string
      enum: [bank_transfer, cash, card, qr, other]
...
    Channel:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        organization_id:
          $ref: "#/components/schemas/UUID"
        kind:
          $ref: "#/components/schemas/ChannelKind"
        name:
          type: string
        is_active:
          type: boolean
      required: [id, organization_id, kind, name, is_active]
    Listing:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        organization_id:
          $ref: "#/components/schemas/UUID"
        unit_id:
          $ref: "#/components/schemas/UUID"
        channel_id:
          $ref: "#/components/schemas/UUID"
        external_listing_id:
          type: string
          nullable: true
        public_name:
          type: string
        ical_import_url:
          type: string
          format: uri
          nullable: true
        is_active:
          type: boolean
      required: [id, organization_id, unit_id, channel_id, public_name, is_active]
...
    Expense:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        organization_id:
          $ref: "#/components/schemas/UUID"
        category:
          $ref: "#/components/schemas/ExpenseCategory"
        expense_date:
          type: string
          format: date
        amount:
          type: number
        currency:
          $ref: "#/components/schemas/Currency"
        payment_method:
          $ref: "#/components/schemas/PaymentMethod"
        invoice_number:
          type: string
          nullable: true
        invoice_ruc:
          type: string
          nullable: true
      required: [id, organization_id, category, expense_date, amount, currency, payment_method]
...
    CreateChannelInput:
      type: object
      properties:
        organization_id:
          $ref: "#/components/schemas/UUID"
        kind:
          $ref: "#/components/schemas/ChannelKind"
        name:
          type: string
      required: [organization_id, kind, name]
    CreateListingInput:
      type: object
      properties:
        organization_id:
          $ref: "#/components/schemas/UUID"
        unit_id:
          $ref: "#/components/schemas/UUID"
        channel_id:
          $ref: "#/components/schemas/UUID"
        external_listing_id:
          type: string
        public_name:
          type: string
        ical_import_url:
          type: string
          format: uri
      required: [organization_id, unit_id, channel_id, public_name]
...
    CreateExpenseInput:
      type: object
      properties:
        organization_id:
          $ref: "#/components/schemas/UUID"
        property_id:
          $ref: "#/components/schemas/UUID"
        unit_id:
          $ref: "#/components/schemas/UUID"
        reservation_id:
          $ref: "#/components/schemas/UUID"
        category:
          $ref: "#/components/schemas/ExpenseCategory"
        vendor_name:
          type: string
        expense_date:
          type: string
          format: date
        amount:
          type: number
        currency:
          $ref: "#/components/schemas/Currency"
        fx_rate_to_pyg:
          type: number
        payment_method:
          $ref: "#/components/schemas/PaymentMethod"
        invoice_number:
          type: string
        invoice_ruc:
          type: string
        receipt_url:
          type: string
          format: uri
        notes:
          type: string
      required: [organization_id, category, expense_date, amount]
...